created a new class Builder, make method build from Builder generic and some modification for inheritance
