+++ design_patterns_project/src/Brands/BMW.java
@@ -4,18 +4,10 @@ import Car.Car;
 
 public class BMW extends Car {
 
-    protected BMW(BMWBuilder builder) {
-        setOwnerName(builder.getOwnerName());
+    public BMW(Builder builder) {
+        super(builder);
+
         setBrand("BMW");
         setPrice(9750);         // starting price for an BMW
-
-        setAutoAirCondition(builder.isAutoAirCondition());
-        setLeatherSeats(builder.isLeatherSeats());
-        setSunRoof(builder.isSunRoof());
-        setBackupCamera(builder.isBackupCamera());
-        setNavigationSystem(builder.isNavigationSystem());
-        setBluetooth(builder.isBluetooth());
-        setRemoteStart(builder.isRemoteStart());
-        setBlindSpotMonitoring(builder.isBlindSpotMonitoring());
     }
 }
