+++ design_patterns_project/src/Brands/BrandsFactory.java
@@ -7,20 +7,20 @@ public class BrandsFactory {
     public static Car getCar(String type, String ownerName) {
 
         if ("BMW-essential".equals(type)) {
-            return new BMWBuilder(ownerName)
-                    .build();
+            return new Builder(ownerName)
+                    .build(BMW.class);
         }
         if ("BMW-comfort".equals(type)) {
-            return new BMWBuilder(ownerName)
+            return new Builder(ownerName)
                     .withBackupCamera()
                     .withNavigationSystem()
                     .withBluetooth()
                     .withBlindSpotMonitoring()
-                    .build();
+                    .build(BMW.class);
         }
 
         if ("BMW-prestige".equals(type)) {
-            return new BMWBuilder(ownerName)
+            return new Builder(ownerName)
                     .withAutoAirCondition()
                     .withLeatherSeats()
                     .withSunRoof()
@@ -29,11 +29,11 @@ public class BrandsFactory {
                     .withBluetooth()
                     .withRemoteStart()
                     .withBlindSpotMonitoring()
-                    .build();
+                    .build(BMW.class);
         }
 
         if ("Audi-prestige".equals(type)) {
-            return new AudiBuilder(ownerName)
+            return new Builder(ownerName)
                     .withAutoAirCondition()
                     .withLeatherSeats()
                     .withSunRoof()
@@ -42,21 +42,21 @@ public class BrandsFactory {
                     .withBluetooth()
                     .withRemoteStart()
                     .withBlindSpotMonitoring()
-                    .build();
+                    .build(Audi.class);
         }
 
         if ("Audi-comfort".equals(type)) {
-            return new AudiBuilder(ownerName)
+            return new Builder(ownerName)
                     .withBackupCamera()
                     .withNavigationSystem()
                     .withBluetooth()
                     .withBlindSpotMonitoring()
-                    .build();
+                    .build(Audi.class);
         }
 
         if ("Audi-essential".equals(type)) {
-            return new AudiBuilder(ownerName)
-                    .build();
+            return new Builder(ownerName)
+                    .build(Audi.class);
         }
 
         return null;
